server/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autopep8==1.7.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
Cython==0.29.33
exceptiongroup==1.1.0
filelock==3.12.0
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.5.0
h11==0.14.0
huggingface-hub==0.14.1
idna==3.4
imageai==3.0.3
iniconfig==2.0.0
itsdangerous==2.1.2
jikan4==0.2.1
jikanpy==4.3.2
Jinja2==3.1.2
kiwisolver==1.4.4
libretranslatepy==2.1.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.7.0
mock==4.0.3
multidict==6.0.4
neat-python==0.92
numpy==1.24.2
oauthlib==3.2.1
openai==0.27.7
opencv-python==4.7.0.72
outcome==1.2.0
packaging==23.0
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.7
pygame==2.3.0
pyparsing==3.0.9
PyPDF2==1.26.0
PySocks==1.7.1
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-translator==1.0.3
PyYAML==6.0
regex==2023.5.5
requests==2.28.1
requests-oauthlib==1.3.1
scipy==1.10.1
selenium==4.8.2
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tk==0.1.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch==1.13.1+cpu
torchvision==0.14.1+cpu
tqdm==4.64.1
transformers==4.29.2
translate==3.6.1
trio==0.22.0
trio-websocket==0.9.2
tweepy==3.8.0
typing_extensions==4.5.0
urllib3==1.26.12
Werkzeug==2.2.2
wsproto==1.2.0
yarl==1.8.2
